separate local sorage, DomEl and projects
