+++ src/JavaScript/projects.js
@@ -1,107 +1,63 @@
+import mobileGlobalDrums from '../myworkImages/mobileGlobalDrummers.png';
+import desktopGlobalDrums from '../myworkImages/desktopGlobalDrummers.png';
+import desktopEcommerce from '../myworkImages/E-commerce product page.png';
+import mobileEcommerce from '../myworkImages/mobileE-commerce-product-page.png';
+
+
 const projects = [
   {
-    name: ' ',
-    popupName: 'Multi Post Stories',
-    description:
-        ' ',
-    featuredImage: 'url(./myworkImages/GlobalDrummers.png)',
-    popupDescription: `Drum Festival Global conference is a yearly drummers community
+    name: 'Global Drummers Festival Website',   
+    description: `Drum Festival Global conference is a yearly drummers community
        event at the global level featuring world-class drummers from more than 100 countries.
         The aim is to provide a single point of Information and registration for all participants. 
         This project was completed with JavaScript, HTML, and CSS.`,
-    popupImage: './images/DesktopPopup.svg',
-    technologies: ['html', 'Css', 'JavaScript'],
-    liveLink: 'https://bobb-rob.github.io/portfolio-project-1/',
-    sourceFile: 'https://github.com/bobb-rob/portfolio-project-1',
+    images: {
+        mobile: mobileGlobalDrums,
+        desktop: `url(${desktopGlobalDrums})`,
+    },   
+    technologies: ['Html', 'CSS', 'JavaScript'],
+    liveLink: 'https://bobb-rob.github.io/Global-Drummers-Workshop/',
+    sourceFile: 'https://github.com/bobb-Rob/Global-Drummers-Workshop',
     id: '1',
   },
   {
-    name: 'Website Portfolio',
-    popupName: 'Multi Post Stories',
+    name: 'Website Portfolio',   
     description:
         "A daily selection of privately personalized reads; no accounts or sign-ups required has been the industry's standard",
-    featuredImage: 'url(./images/project-card-img.svg)',
-    popupDescription: `Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the 
-        industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to 
-        make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent`,
-    popupImage: './images/DesktopPopup.svg',
-    technologies: ['html', 'bootstrap', 'ruby and rails'],
+    images: {
+        mobile: mobileGlobalDrums,
+        desktop: `url(${desktopGlobalDrums})`,
+    },    
+    technologies: ['Html', 'CSS', 'JavaScript'],
     liveLink: 'https://bobb-rob.github.io/portfolio-project-1/',
     sourceFile: 'https://github.com/bobb-rob/portfolio-project-1',
     id: '2',
   },
   {
-    name: 'Profesional Art Printing Data',
-    popupName: 'Multi Post Stories',
+    name: 'Profesional Art Printing Data',  
     description:
         "A daily selection of privately personalized reads; no accounts or sign-ups required has been the industry's standard",
-    featuredImage: 'url(./images/project-card-img.svg)',
-    popupDescription: `Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the 
-        industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to 
-        make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent`,
-    popupImage: './images/DesktopPopup.svg',
-    technologies: ['html', 'bootstrap', 'ruby and rails'],
+    images: {
+        mobile: mobileGlobalDrums,
+        desktop: `url(${desktopGlobalDrums})`,
+    },
+    technologies:  ['Html', 'CSS', 'JavaScript'],
     liveLink: 'https://bobb-rob.github.io/portfolio-project-1/',
     sourceFile: 'https://github.com/bobb-rob/portfolio-project-1',
     id: '3',
-  },
-  {
-    name: 'Profesional Art Printing Data',
-    popupName: 'Multi Post Stories',
-    description:
-        "A daily selection of privately personalized reads; no accounts or sign-ups required has been the industry's standard",
-    featuredImage: 'url(./images/project-card-img.svg)',
-    popupDescription: `Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the 
-        industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to 
-        make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent`,
-    popupImage: './images/DesktopPopup.svg',
-    technologies: ['html', 'bootstrap', 'ruby and rails'],
-    liveLink: 'https://bobb-rob.github.io/portfolio-project-1/',
-    sourceFile: 'https://github.com/bobb-rob/portfolio-project-1',
-    id: '4',
-  },
-  {
-    name: 'Profesional Art Printing Data',
-    popupName: 'Multi Post Stories',
-    description:
-        "A daily selection of privately personalized reads; no accounts or sign-ups required has been the industry's standard",
-    featuredImage: 'url(./images/project-card-img.svg)',
-    popupDescription: `Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the 
-        industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to 
-        make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent`,
-    popupImage: './images/DesktopPopup.svg',
-    technologies: ['html', 'bootstrap', 'ruby and rails'],
-    liveLink: 'https://bobb-rob.github.io/portfolio-project-1/',
-    sourceFile: 'https://github.com/bobb-rob/portfolio-project-1',
-    id: '5',
-  },
-  {
-    name: 'Profesional Art Printing Data',
-    popupName: 'Multi Post Stories',
-    description:
-        "A daily selection of privately personalized reads; no accounts or sign-ups required has been the industry's standard",
-    featuredImage: 'url(./images/project-card-img.svg)',
-    popupDescription: `Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the 
-        industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to 
-        make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent`,
-    popupImage: './images/DesktopPopup.svg',
-    technologies: ['html', 'bootstrap', 'ruby and rails'],
-    liveLink: 'https://bobb-rob.github.io/portfolio-project-1/',
-    sourceFile: 'https://github.com/bobb-rob/portfolio-project-1',
-    id: '6',
-  },
+  },  
 ];
 
 const projects2 = [{
-  name: 'Profesional Art Printing Data',
-  popupName: 'E-commerce product page',
+  name: 'E-commerce product page', 
   description:
-      "A daily selection of privately personalized reads; no accounts or sign-ups required has been the industry's standard",
-  featuredImage: 'url(./images/g-festiv.png)',
-  popupDescription: `This is a responsive single page e-commerce product page, a solution to the E-commerce product page challenge on Frontend Mentor. 
-    Frontend Mentor challenges help you improve your coding skills by building realistic projects.`,
-  popupImage: './myworkImages/E-commerce product page.png',
-  technologies: ['HTML', 'CSS', 'JavaScript'],
+  `This is a responsive single page e-commerce product page, a solution to the E-commerce product page challenge on Frontend Mentor. 
+  Frontend Mentor challenges help you improve your coding skills by building realistic projects.`,
+  images: {
+    mobile: mobileEcommerce,
+    desktop:  `url(${desktopEcommerce})`,
+  },  
+  technologies: ['Html', 'CSS', 'JavaScript'],
   liveLink: 'https://bobb-rob.github.io/ecommerce-product-page/',
   sourceFile: 'https://github.com/bobb-Rob/ecommerce-product-page',
   id: 'x1',
