+++ src/index.js
@@ -1,141 +1,13 @@
 import { projects, projects2 } from './JavaScript/projects.js';
+import DOM from './JavaScript/domEl.js';
+import setLocalStore from './JavaScript/localStorage.js';
 import displayImages from './JavaScript/images.js';
 import './css/style.css';
 
 displayImages();
 
-const setLocalStore = (() => {
-  function useStorageInfo() {
-    const currentInfo = JSON.parse(localStorage.getItem('contactInfo'));
+DOM.displayProjects(projects);
 
-    document.getElementById('visitors-name').value = currentInfo.fullName;
-    document.getElementById('visitors-email').value = currentInfo.email;
-    document.getElementById('contact-text-area').value = currentInfo.message;
-  }
-
-  function setStorage() {
-    const contactInfo = {
-      fullName: document.getElementById('visitors-name').value,
-      email: document.getElementById('visitors-email').value,
-      message: document.getElementById('contact-text-area').value,
-    };
-    localStorage.setItem('contactInfo', JSON.stringify(contactInfo));
-    useStorageInfo();
-  }
-
-  return { setStorage, useStorageInfo };
-})();
-
-const DOM = (() => {
-  // create Element function
-  function createElement(elem, className = null) {
-    const element = document.createElement(elem);
-    if (className !== null) {
-      element.classList.add(className);
-    }
-    return element;
-  }
-
-  // create close icon function
-  const createCloseIcon = (location, closeWrap) => {
-    const closeIcon = createElement('img', 'close-icon');
-    const closeWrapper = createElement(closeWrap, 'close-icon-wrapper');
-    closeIcon.src = './icons/close-icon.png';
-    closeIcon.alt = 'close icon';
-    closeWrapper.appendChild(closeIcon);
-    location.appendChild(closeWrapper);
-  };
-
-  // Projects card display function
-  const cardContainer = document.getElementById('card-container');
-  const createProjectCard = (project) => {
-    const listItem = createElement('li', 'project-card');
-    listItem.style.background = `${project.featuredImage}`;
-    listItem.id = project.id;
-
-    const projectName = createElement('h3');
-    projectName.textContent = `${project.name}`;
-
-    const projectDescription = createElement('p');
-    projectDescription.textContent = `${project.description}`;
-
-    const techListItem1 = createElement('li');
-    techListItem1.textContent = `${project.technologies[0]}`;
-    const techListItem2 = createElement('li');
-    techListItem2.textContent = `${project.technologies[1]}`;
-    const techListItem3 = createElement('li');
-    techListItem3.textContent = `${project.technologies[2]}`;
-    const techUL = createElement('ul', 'description-tags');
-    techUL.append(techListItem1, techListItem2, techListItem3);
-
-    const cardDescription = createElement('div', 'card-description');
-    cardDescription.appendChild(projectName);
-    cardDescription.appendChild(projectDescription);
-    cardDescription.appendChild(techUL);
-
-    const projectBtn = createElement('button', 'btn-card');
-    projectBtn.classList.add('btn');
-    projectBtn.type = 'button';
-    projectBtn.textContent = 'See Project';
-    listItem.appendChild(cardDescription);
-    listItem.appendChild(projectBtn);
-    return listItem;
-  };
-
-  // Project card display map
-  projects.map((project) => cardContainer.appendChild(createProjectCard(project)));
-
-  // Project pop up window create function
-  const createPopupWindow = (project) => {
-    const projectModal = createElement('div', 'project-modal');
-    const popupContainer = createElement('div', 'popup-container');
-    popupContainer.innerHTML = `
-            <h2>${project.popupName}</h2>
-            <ul class="tech-stack">
-              ${project.technologies.map((tech) => `<li>${tech}</li>`)}
-            </ul>
-            <div class="popup-img-p">
-                <img src="${project.popupImage}" alt="${project.name}">
-
-                <div class="popup-btn-wrapper">
-                    <p>${project.popupDescription}</p>
-                    
-                    <div class="pop-btn-wrap">
-                      <button class="btn popup-btn" type="button">
-                          <a href="${project.liveLink}">See Live
-                          <img src="./icons/new-window-icon.svg" alt="New window icon">
-                          </a>                        
-                      </button>
-                      <button class="btn popup-btn" type="button">
-                          <a href="${project.sourceFile}">See Source
-                          <img src="./icons/github-icon.svg" alt="github icon">
-                          </a>                        
-                      </button>
-                    </div>
-                </div>
-            </div>
-        `;
-    projectModal.appendChild(popupContainer);
-    return projectModal;
-  };
-
-  // elements
-  const hamburgerBtn = document.querySelector('.hamburger-btn'); // Hamburger
-  const navLinksWrapper = document.getElementById('nav-links'); // Menu link wrapper
-  const projectBtns = document.querySelectorAll('.btn-card'); // See project btns
-  const workSection = document.getElementById('recent-works'); // portfolio Work section
-
-  return {
-    createElement,
-    createCloseIcon,
-    createPopupWindow,
-    hamburgerBtn,
-    navLinksWrapper,
-    cardContainer,
-    projectBtns,
-    workSection,
-  };
-})();
 
 const formValidation = () => {
   // Email lowercase sensitive validation
